moved user interface to the bin
